### 2022-05-30 481bf70
Include DOIs in references
