+++ paper/paper.bib
@@ -7,7 +7,8 @@
   number={170},
   pages={269--279},
   year={2020},
-  publisher={Academia Colombiana de Ciencias Exactas, F{\'\i}sicas y Naturales}
+  publisher={Academia Colombiana de Ciencias Exactas, F{\'\i}sicas y Naturales}, 
+  doi="10.18257/raccefyn.960"
 }
 
 @article{10EW,
@@ -18,7 +19,8 @@
   number={2},
   pages={024501},
   year={2010},
-  publisher={APS}
+  publisher={APS},
+  doi="10.1103/physreva.82.024501"
 }
 
 @article{14NW,
@@ -29,7 +31,8 @@
   number={36},
   pages={7623--7639},
   year={2014},
-  publisher={ACS Publications}
+  publisher={ACS Publications},
+  doi="10.1021/jp504058s",
 }
 
 @article{17NW,
@@ -40,7 +43,8 @@
   number={2},
   pages={577--586},
   year={2017},
-  publisher={ACS Publications}
+  publisher={ACS Publications},
+  doi="10.1021/acs.jctc.6b01050"
 }
 
 @article{18JW,
@@ -51,7 +55,8 @@
   number={16},
   pages={164112},
   year={2018},
-  publisher={AIP Publishing LLC}
+  publisher={AIP Publishing LLC},
+  doi="10.1063/1.5051455"
 }
 
 @article{18JW2,
@@ -62,7 +67,8 @@
   number={1},
   pages={e25425},
   year={2018},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi="10.1002/qua.25425"
 }
 
 @article{22SW,
@@ -71,7 +77,8 @@
   journal={Wiley Interdisciplinary Reviews: Computational Molecular Science},
   pages={e1617},
   year={2022},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi="10.1002/wcms.1617"
 }
 
 % DFT
@@ -83,7 +90,8 @@
   number={5890},
   pages={792--794},
   year={2008},
-  publisher={American Association for the Advancement of Science}
+  publisher={American Association for the Advancement of Science},
+  doi="10.1126/science.1158722"
 }
 
 @article{03TS,
@@ -94,7 +102,8 @@
   number={14},
   pages={146401},
   year={2003},
-  publisher={APS}
+  publisher={APS},
+  doi="10.1103/physrevlett.91.146401"
 }
 
 @article{libxc,
@@ -105,7 +114,8 @@
   number={10},
   pages={2272--2281},
   year={2012},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi="10.1016/j.cpc.2012.05.007"
 }
 
 % BOOOKS __________________________________________________
@@ -120,7 +130,8 @@
   title={Mathematical methods for physicists},
   author={Arfken, George B and Weber, Hans J},
   year={1999},
-  publisher={American Association of Physics Teachers}
+  publisher={American Association of Physics Teachers},
+  doi="10.1016/c2009\{-}1\{-}28537\{-}6",
 }
 
 % BASIS SETS vs grid
@@ -132,7 +143,8 @@
   number={1},
   pages={21--34},
   year={2013},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi="10.1002/qua.24355"
 }
 
 @article{octopus,
@@ -143,7 +155,8 @@
   number={47},
   pages={31371--31396},
   year={2015},
-  publisher={Royal Society of Chemistry}
+  publisher={Royal Society of Chemistry},
+  doi="10.1039/C5CP00351B"
 }
 
 @article{17RS,
[…]
