### 2022-05-30 #4426 @VHchavez
Hello @arfon, hello @Kevin-Mattheus-Moerman. Thank you for your editorial work.
Thank you, @lucydot for your review 😸

1. The references are now updated. I have been unable to find a DOI for one of the books. What is the appropriate way to proceed?
2. The affiliation for my entry is failing. I have checked for correctness, but I have been unable to find any problem/typo. Any help is very much appreciated.
3. I believe @ raghurama123 (from your database) would be quite suited for this work.

Thanks again!
