### 2022-06-09 2335ab4
Minor corrections and a suggestion
