### 2022-06-09 4261f3d
Fix Introduction, typos and a suggestion
Please address the comment `[COMMENT raghurama123]: This is a bit confusing. A reader may wonder if a diatomic with heavy atoms (say Uranium dimer) qualifies as a system with a handful of electrons. Please remove this sentence and/or include another point`
