+++ paper/paper.md
@@ -19,23 +19,27 @@ affiliations:
       index: 1  
     - name: Gridspace, Los Angeles, California, USA  
       index: 2  
-    - name: Department of Physics and Astronomy, Purdue Unversity, West Lafayette, Indiana, USA  
+    - name: Department of Physics and Astronomy, Purdue University, West Lafayette, Indiana, USA  
       index: 3  
 date: 16 Mayo 2022  
 bibliography: paper.bib  
 
 ---
+# Introduction
 
-# Summary 
 Diatomic molecules are among the most useful systems to test new ideas in quantum chemistry:
 
-- With only a handful of electrons, their computing time is short.
-- They highlight the present and ubiquitous problems of modern approximations.
+- With only a handful of electrons, their computing time is short. `[COMMENT raghurama123]: This is a bit confusing. A reader may wonder if a diatomic with heavy atoms (say Uranium dimer) qualifies as a system with a handful of electrons. Please remove this sentence and/or include another point` 
+- They highlight the present and ubiquitous problems of modern approximations. 
 - By defining two fragments with an atom each, diatomics are ideal to implement and test quantum embedding methods.
 
-In this work, we introduce ``pyCADMium``, a Python module that uses a prolate spheroidal (PS) coordinate grid to accurately perform computational chemistry calculations on on systems with cylindrical symmetry. The name is an acronym for ``Chemical Atoms in Diatomic Molecules".``pyCADMium`` originated in a proprietary programming language but has been rewritten from the ground up as an open-source alternative. The code has been the main driving force behind the development of "Partition Density Functional Theory," [@20CW;@10EW;@17NW;@18JW] a method that uses quantum embedding to lower the cost of a calculation and fixes problems inherent to density functional approximations [@08CY].  
+Most practical calculations use a finite basis set of fixed functional form to represent operators and quantities like potentials, orbitals, and densities [@12SO;@13H]. Such approaches require analytic expressions for the matrix representation of various operators and often exhibit poor basis set convergence.  The basis set incompleteness error can be minimized by increasing the number of basis functions used, but it cannot be entirely eliminated in practice because of the lack of parameters (for instance, exponents, contraction coefficients, etc.) for large basis sets. On the other hand, grid-based methods intrinsically allow for an accurate representation of operators and permit the use of arbitrarily large basis sets [@octopus]. Nevertheless, the number of points required to achieve a significant accuracy can quickly become unmanageable. ``pyCADMium`` presented in this work, uses a prolate spheroidal (PS) grid to circumvent these issues for atoms and diatomic molecules. 
 
-Most practical calculations use a basis-set to represent operators and quantities like potentials, orbitals, and densities [@12SO;@13H]. The lack of an accurate space representation of these and other operators gives rise to the basis-set incompleteness error. This error can be minimized by increasing the number of basis functions used, but it cannot be entirely eliminated in practice. On the other hand, grid-based methods intrinsically allow for an accurate representation [@octopus]. Nevertheless, the number of points required to achieve a significant accuracy can become quickly unmanageable. ``pyCADMium`` uses a PS grid to circumvent these issues. The PS grid is a coordinate system formed by revolving an elliptic coordinate plane around the line intersecting the two foci. These planes are formed by ellipses and hyperbolae that share the same focus [@arfken]. Atoms and diatomics are ideally suited for this coordinate system given that each foci can be used to allocate an atom. Additionally, the PS grid is denser around the foci in its cartesian representation, where we expect functions of molecular systems to change rapidly [@17RS].  
+# Summary 
+
[…]
