### 2022-06-09 #4459 @editorialbot
[REVIEW]: pyCADMium: Chemical Atoms in Diatomic Molecules. A prolate spheroidal Python module for embedding calculations.
**Submitting author:** @VHchavez (<a href="http://orcid.org/0000-0003-3765-2961">Victor Hugo Gonzalez Chavez</a>)
**Repository:** https://github.com/wasserman-group/pyCADMium
**Branch with paper.md** (empty if default branch): main
**Version:** v0.9.1
