### 2022-06-09 #4459 @raghurama123
Here are some initial comments on the content of Readme.md and paper.md: [pull request](https://github.com/wasserman-group/pyCADMium/pull/9). In each document, I have a comment in single quotes beginning with `[COMMENT raghurama123]: ...`. Please address these and remove my comments.

I will provide the software review soon.
