### 2022-06-19 wasserman-group/pyCADMium#10 @srmnitc
JOSS review issues
This is part of the review for JOSS at https://github.com/openjournals/joss-reviews/issues/4459 .

The following issues were found during the review:

## Functionality
### Installation
- Installation instructions are given on the README, but not on documentation page.
- Installation instructions for `libxc`: The `tar` command is given in the wrong order. It should be `tar` first, then `cd` command. Also, the `wget` command download a file called `down.php?file=5.0.0%2Flibxc-5.0.0.tar.gz`.

### Functionality
- Please see https://github.com/wasserman-group/pyCADMium_examples/issues/3
- Functionality is not clear from the documentation page. The available notebooks can be easily added to the existing documentation.

## Documentation
### A Statement of need
- No statement of need in documentation.
### Installation instructions
- No installation instructions in the documentation
### Example usage
- Examples available in a separate repository, but not in documentation
### Functionality documentation
- Class methods are, for example, in `kohnsham.py`, have no docstrings. When docstrings are present, the arguments and return values are not specified
### Automated tests
- Automated tests exist, but it is not explained anywhere on how to run them properly
- Tests fail because there are packages that are not installed with the installation instructions, but are needed. Eg- `pytest`, `pydantic`
- Not super important, but tests raises a warning with the statement `import numpy.matlib as matlab`
### Community guidelines
- Apart from an email id to contact in case of errors, no other infomation is available.
## Software paper
### Quality of writing
- There are some typos, for example in lines 45, 53, 54, 67. A spellcheck of the whole paper will fix these issues
- lines 10-13 needs a citation for it to be helpful for the general audience
- lines 26-27 "This error can be......eliminated in practice" also needs a citation in my opinion
- line 52, a citation could be helpful
- line 61 acronym P-DFT was not introduced before
