### 2022-06-19 #4459 @srmnitc
@lucydot @VHchavez I have completed my first review of the package. I think pyCADMium is a useful package that should be published in JOSS. The code is well-written, and the set of examples are convincing. The JOSS paper is also well written. There are some issues which prevent me from checking all the boxes. Most of them are related to documentation and functionality. I have raised two separate issues which provide a detailed description of the issues (https://github.com/wasserman-group/pyCADMium/issues/10, https://github.com/wasserman-group/pyCADMium_examples/issues/3).

Functionality: There are a number of useful examples. However, the dependencies for running these examples are not correctly specified and some of them run into errors. Also, including these examples as part of the documentation will greatly improve the utility of the documentation.

Documentation: Documentation is split in the README and the github pages. Github pages documentation is very well written, but could benefit from the addition of some of the things that I mentioned in the issue. The module comes with a full set of automated tests, documenting how to run them will make it much easier for the user.

Overall, the package is useful and deserve publication. I will be happy to check the remaining boxes when the issues are addressed. Thanks again for the nice work @VHchavez .
