### 2022-06-21 ceee3f9
Update paper.md
Update paper submission based on @raghurama123 recommendations. 
DARSEC comparison included in an incoming commit.
