### 2022-07-18 wasserman-group/pyCADMium#10 @VHchavez
Hello Sarath,

Thanks for your much helpful comments. I am currently working on your edits on the fly. Otherwise, I find it harder to do it all in one go. I am editing from top to bottom.

- I have added an analogous installation page to the README in the documentation.
- I have updated the pylibxc instructions using the much more efficient `pylibxc2` mentioned in the review.
- I tried the `wget` command and I do get a "clean" download file. Nevertheless, I added a comment so that users are careful in that regard.

For the functionality, is it enough to add a page that explains the examples as well as links to the example repository, or are you suggesting each of them be added within the documentation. In that case, would adding a block of code suffice? I know that there are smarter ways of embedding `.ipynb` but I am unsure how to do it.
