### 2022-07-19 wasserman-group/pyCADMium#10 @srmnitc
@VHchavez [nbsphinx](https://nbsphinx.readthedocs.io/en/0.8.9/) can help you embed the notebooks directly. But I think a description of the examples with corresponding links would also be helpful!
