### 2022-08-13 7efeaf1
Add sentence to community guidelines
