### 2022-08-13 f702cc3
Add pydantic, pytest to required libraries
