### 2022-08-13 bcd8310
Exit instead of raising error for orbital invert function
