### 2022-08-13 ec18700
Add missing docgstrings
