+++ .github/workflows/CI.yml
@@ -12,7 +12,7 @@ jobs:
         # - conda-env: test_env
         #   python-version: 3.6
         - conda-env: test_env
-          python-version: 3.7
+          python-version: 3.8
         # - conda-env: test_env
         #   python-version: 3.8
 
