+++ docs/examples/INV/README.md
@@ -1,0 +1,17 @@
+### Density-Based Partitioning Methods for Ground-State Molecular Calculations
+#### Jonathan Nafziger and Adam Wasserman
+
+With the growing complexity of systems that can be treated with modern
+electronic-structure methods, it is critical to develop accurate and efficient strategies to
+partition the systems into smaller, more tractable fragments. We review some of the various
+recent formalisms that have been proposed to achieve this goal using fragment (ground-state)
+electron densities as the main variables, with an emphasis on partition density-functional
+theory (PDFT), which the authors have been developing. To expose the subtle but important
+differences between alternative approaches and to highlight the challenges involved with
+density partitioning, we focus on the simplest possible systems where the various methods can
+be transparently compared. We provide benchmark PDFT calculations on homonuclear
+diatomic molecules and analyze the associated partition potentials. We derive a new exact
+condition determining the strength of the singularities of the partition potentials at the nuclei,
+establish the connection between charge-transfer and electronegativity equalization between
+fragments, test different ways of dealing with fractional fragment charges and spins, and finally outline a general strategy for
+overcoming delocalization and static-correlation errors in density-functional calculations.
