+++ docs/index.rst
@@ -11,14 +11,79 @@ Python module that uses a prolate spheroidal (PS) coordinate grid to accurately 
 
 
 .. toctree::
-   :maxdepth: 3
-   :caption: Contents:
+   :maxdepth: 1
+   :caption: CONTENTS:
 
    installation
    coordinate_system
    getting_started
    api
+   
 
+.. toctree::
+   :maxdepth: 1
+   :caption: GRID EXAMPLES:
+
+   examples/Grids/README.md
+   examples/Grids/Grid.ipynb
+   examples/Grids/CADMIUM_LOGO.ipynb
+
+
+.. toctree::
+   :maxdepth: 1
+   :caption: KOHN-SHAM DFT ATOM EXAMPLES:
+
+   examples/KS_DFT/00_README.ipynb
+   examples/KS_DFT/01_H_Exact.ipynb
+   examples/KS_DFT/02_LDA_H.ipynb
+   examples/KS_DFT/03_LDA_He.ipynb
+   examples/KS_DFT/04_LDA_Li.ipynb
+   examples/KS_DFT/05_LDA_Be.ipynb
+   examples/KS_DFT/06_LDA_B.ipynb
+   examples/KS_DFT/07_LDA_C.ipynb
+   examples/KS_DFT/08_LDA_N.ipynb
+   examples/KS_DFT/09_LDA_O.ipynb
+   examples/KS_DFT/10_LDA_F.ipynb
+   examples/KS_DFT/11_LDA_Ne.ipynb
+
+.. toctree::
+   :maxdepth: 1
+   :caption: KOHN-SHAM DFT DIATOMIC EXAMPLES:
+
+   examples/KS_DFT/00_README2.ipynb
+   examples/KS_DFT/01_NaCl.ipynb
+   examples/KS_DFT/02_CO.ipynb
+   examples/KS_DFT/03_O2.ipynb
+   examples/KS_DFT/04_F2.ipynb
+
+.. toctree::
+   :maxdepth: 1
+   :caption: PDFT INVERSION EXAMPLES:
+
+   examples/INV/README.ipynb
+   examples/INV/PDFT_Calculation_Be2_Orbital.ipynb
+   examples/INV/PDFT_Calculation_H2plus.ipynb
+   examples/INV/PDFT_Calculations_H2.ipynb
+   examples/INV/PDFT_Calculations_He2.ipynb
+   examples/INV/PDFT_Calculations_Li2.ipynb
+   examples/INV/ENS_vs_FOO/H2_ens.ipynb
+   examples/INV/ENS_vs_FOO/H2_foo.ipynb
+   examples/INV/ENS_vs_FOO/h2plus_ens.ipynb
+   examples/INV/ENS_vs_FOO/h2plus_foo.ipynb
+   examples/INV/ENS_vs_FOO/Li2_ens.ipynb
+   examples/INV/ENS_vs_FOO/Li2_foo.ipynb
+
+.. toctree::
+   :maxdepth: 1
+   :caption: DELOCALIZATION AND STATIC CORRELATION EXAMPLES:
+
+   examples/STRETCHING/README.ipynb
+   examples/STRETCHING/Figure_1.ipynb
+   examples/STRETCHING/H2_Stretched.ipynb
+   examples/STRETCHING/H2_Stretched_LDA.ipynb
+   examples/STRETCHING/H2plus_Stretched.ipynb
+   examples/STRETCHING/H2plus_Stretched_LDA.ipynb
+   examples/STRETCHING/Li2.ipynb
 
 
 Indices and tables
