### 2022-08-15 wasserman-group/pyCADMium#10 @VHchavez
Hello Sarath,

I've followed your advice on the documentation and added each of the examples as well. I had problems adding `.md` files for some weird reason. I spied on several repositories and I noticed that people add them like normal `.rts`. But I could never figure that out.

Added the missing information, like the installation to the documentation.

The KohnSham class now has docstrings to accompany each of its methods.

Pytest details have been added to the repo's README. And all of the dependencies have been added to the `setup.py`

I encouraged folks to submit issues and pull requests. I am not sure what else can be added.

Thanks for reviewing the document with the huge amount of spelling errors 🙈. They have been addressed as well as the sentences with no citation.
