+++ paper/paper.bib
@@ -118,6 +118,28 @@
   doi="10.1016/j.cpc.2012.05.007"
 }
 
+@article{darsec,
+  title={Fully numerical all-electron solutions of the optimized effective potential equation for diatomic molecules},
+  author={Makmal, Adi and Kummel, Stephan and Kronik, Leeor},
+  journal={Journal of chemical theory and computation},
+  volume={5},
+  number={7},
+  pages={1731--1740},
+  year={2009},
+  publisher={ACS Publications}
+}
+
+@article{03WY,
+  title={A direct optimization method for calculating density functionals and exchange--correlation potentials from electron densities},
+  author={Wu, Qin and Yang, Weitao},
+  journal={The Journal of chemical physics},
+  volume={118},
+  number={6},
+  pages={2498--2509},
+  year={2003},
+  publisher={American Institute of Physics}
+}
+
 % BOOOKS __________________________________________________
 @book{12SO,
   title={Modern quantum chemistry: introduction to advanced electronic structure theory},
