+++ paper/paper.md
@@ -30,13 +30,13 @@ bibliography: paper.bib
 Diatomic molecules are among the most useful systems to test new ideas in quantum chemistry:
 
 - They highlight the present and ubiquitous problems of modern approximations [@08CY]. 
-- By defining two fragments with an atom each, diatomics are ideal to implement and test quantum embedding methods [20CW].
+- By defining two fragments with an atom each, diatomics are ideal to implement and test quantum embedding methods [@20CW].
 
 Most practical calculations use a finite basis set of fixed functional form to represent operators and quantities like potentials, orbitals, and densities [@12SO;@13H]. Such approaches require analytic expressions for the matrix representation of various operators and often exhibit poor basis set convergence.  The basis set incompleteness error can be minimized by increasing the number of basis functions used, but it cannot be entirely eliminated in practice because of the lack of parameters (for instance, exponents, contraction coefficients, etc.) for large basis sets. On the other hand, grid-based methods intrinsically allow for an accurate representation of operators and permit the use of arbitrarily large basis sets [@octopus]. Nevertheless, the number of points required to achieve a significant accuracy can quickly become unmanageable. ``pyCADMium`` presented in this work, uses a prolate spheroidal (PS) grid to circumvent these issues for atoms and diatomic molecules. 
 
 # Summary 
 
-In this work, we introduce ``pyCADMium``, a Python module that uses a PS coordinate grid to accurately perform computational chemistry calculations on systems with cylindrical symmetry. The name is an acronym for ``Chemical Atoms in Diatomic Molecules``.``pyCADMium`` originated in a proprietary programming language but has been rewritten from the ground up as an open-source alternative. The code has been the main driving force behind the development of "Partition Density Functional Theory," [@20CW;@10EW;@17NW;@18JW] a method that uses quantum embedding to lower the cost of a calculation and fixes problems inherent to density functional approximations [@08CY].  
+In this work, we introduce ``pyCADMium``, a Python module that uses a PS coordinate grid to accurately perform computational chemistry calculations on systems with cylindrical symmetry. The name is an acronym for ``Chemical Atoms in Diatomic Molecules``. ``pyCADMium`` originated in a proprietary programming language but has been rewritten from the ground up as an open-source alternative. The code has been the main driving force behind the development of "Partition Density Functional Theory," [@20CW;@10EW;@17NW;@18JW] a method that uses quantum embedding to lower the cost of a calculation and fixes problems inherent to density functional approximations [@08CY].  
 
 The PS grid is a coordinate system formed by revolving an elliptic coordinate plane around the line intersecting the two foci. These planes are formed by ellipses and hyperbolae that share the same focus [@arfken]. Atoms and diatomics are ideally suited for this coordinate system given that each foci can be used to allocate an atom. Additionally, the PS grid is denser around the foci in its cartesian representation, where we expect wave functions of molecular systems to change rapidly [@17RS].  
 
@@ -48,15 +48,15 @@ Once the fragments and/or molecule is defined. ``pyCADMium`` can perform:
 
 1. Density functional theory (DFT) calculation. Choose a density functional approximation up to the generalized gradient approximation (GGA) rung [@03TS] from the library of exchange correlation functionals, libxc [@libxc], and perform a self-consistent Kohn-Sham DFT calculation to find the energies, orbitals and density.  
 
-2. Density-to-Potential inversion. Given any density on the PS grid, perform a numerical inversion [@22SW] to find the multiplicative external potential that reproduces the input density[03WY]. This problem is ill-posed when the potential is expressed on a basis-set, but it is well-posed when done on a grid [@18JW2].  
[…]
