### 2022-08-22 #4459 @lucydot
@srmnitc @raghurama123 looks like it is back to you - @VHchavez has responded to the points raised over at the issue - https://github.com/wasserman-group/pyCADMium/issues/10.
