### 2022-08-28 #4459 @raghurama123
I am happy with the present version of the software paper. I would like to add that I have performed some calculations with pyCADMium and I was able to figure out how to make any modifications to the source. However, I have not explored all possible options in the code.

The authors have addressed all my previous remarks and included a citation to the DARSEC code. As a difference between the implementations in pyCADMium and DARSEC, the authors have clarified that pyCADMium uses libxc while DARSEC presumably uses XC functionals coded up separately. I could not find elsewhere if DARSEC also uses libxc. Since pyCADMium provides the source code, other researchers can play with the code and perform calculations in prolate spheroidal coordinates.

Here are a  couple of small typos in the paper, which the authors may want to fix.
line-43: `Place one (A, )` does not make sense. Maybe this can be changed to `Place one atom (A) at the origin`?
line-46: Remove period after defined

Once these typos are fixed, the article and the code may be accepted without any further review.
