### 2022-08-28 ff34897
Update paper.md
Fix Typo
