### 2022-08-28 #4459 @VHchavez
Thanks for your comments, @raghurama123 :)

In regards to the comments on line 43. The initial idea was to add a ∅ to signify that the site was empty, but I believe it gets lost in the process of generating the pdf. If I'm unable to render it appropriately, do you think (A, 0) would be clear enough? Otherwise, I can leave it as (A).
I like the idea of having the two different sites explicit since the grid *is* generated independently of whether there's an atom or not.

Line 46 has been updated.
