### 2022-08-28 #4459 @raghurama123
Okay, that makes sense. I was confused about the notation. Origin shouldn't be the right choice in this coordinate system as the density of grids may not be good there. Perhaps you can change line 43 to `Place one atom at a focus (A)` ?
