### 2022-08-28 615ec76
Update paper.md
Fix typo
