### 2022-08-31 #4459 @lucydot
Once you have done this, please:

- Make a tagged release of your software
- Archive this release on Zenodo (or similar)
- **The archive should have exactly the same title and author list as your JOSS paper**

Report the archive DOI and version number in this thread 👇
