+++ paper/paper.bib
@@ -126,7 +126,8 @@
   number={7},
   pages={1731--1740},
   year={2009},
-  publisher={ACS Publications}
+  publisher={ACS Publications}, 
+  doi="10.1021/ct800485v"
 }
 
 @article{03WY,
@@ -137,7 +138,8 @@
   number={6},
   pages={2498--2509},
   year={2003},
-  publisher={American Institute of Physics}
+  publisher={American Institute of Physics},
+  doi="10.1063/1.1535422"
 }
 
 % BOOOKS __________________________________________________
