### 2022-09-01 26b4751
Revert "Add missing DOI to references"
This reverts commit 3814c29c60d03ce44070893c2b4c25ca66b6dc81.
