+++ paper/paper.bib
@@ -126,8 +126,7 @@
   number={7},
   pages={1731--1740},
   year={2009},
-  publisher={ACS Publications}, 
-  doi="10.1021/ct800485v"
+  publisher={ACS Publications}
 }
 
 @article{03WY,
@@ -138,8 +137,7 @@
   number={6},
   pages={2498--2509},
   year={2003},
-  publisher={American Institute of Physics},
-  doi="10.1063/1.1535422"
+  publisher={American Institute of Physics}
 }
 
 % BOOOKS __________________________________________________
