### 2022-09-05 #4459 @lucydot
@VHchavez  - the latest paper doesn't include the changes you made - it looks like your commit which included the additional dois has been reverted? (https://github.com/wasserman-group/pyCADMium/commit/26b4751229e412049f4f968ab3d475ff6fd0de9d)
