+++ paper/paper.bib
@@ -196,7 +196,7 @@
 }
 
 @article{B82,
-  title={Numerical Hartree--Fock--Slater calculations on diatomic molecules},
+  title={Numerical {H}artree--{F}ock--{S}later calculations on diatomic molecules},
   author={Becke, Axel Dieter},
   journal={The Journal of Chemical Physics},
   volume={76},
