### 2022-09-12 #4459 @danielskatz
@VHchavez - I'm the AEiC who will finalize this.  I see a few changes needed, as shown in https://github.com/wasserman-group/pyCADMium/pull/11 - please merge this, or let me know what you disagree with, then we can continue to acceptance and publication.
