### 2022-09-19 #4459 @susilehtola
The current reference to Libxc is not Marques et al 2012 but

> S. Lehtola, C. Steigemann, M. J. T. Oliveira, and M. A. L. Marques, Recent developments in LIBXC — a comprehensive library of functionals for density functional theory, SoftwareX 7, 1 (2018). doi:10.1016/j.softx.2017.11.002

Fully numerical calculations on atoms and diatomic molecules have been recently reviewed in

> S. Lehtola, A review on non-relativistic fully numerical electronic structure calculations on atoms and diatomic molecules, Int. J. Quantum Chem. 119, e25968 (2019). doi:10.1002/qua.25968

An open-source implementation of the finite element method for diatomic molecules in the prolate spheroidal coordinate system has been described in

> S. Lehtola, Fully numerical Hartree–Fock and density functional calculations. II. Diatomic molecules, Int. J. Quantum Chem. 119, e25944 (2019). doi:10.1002/qua.25944

Note that HelFEM is open source, as well, and includes a full interface to Libxc.

I also I do not see any writeup of the algorithms or method of solution used in the code.
