### 2022-09-22 #4459 @choochoo-bot
Thanks @susilehtola I'll ask other editors if it is possible to update the Libxc reference somehow.

RE: algorithms,methods of solutions: I don't know if you have seen the more extensive documentation online - perhaps not? (https://wasserman-group.github.io/pyCADMium/coordinate_system.html). A JOSS paper is designed to give a more high-level overview of functionality rather than dig into the details of underlying methods.
